rename: inGame 관련 컴포넌트 폴더 구조 수정(waiting/game/공용 구분)
