+++ packages/client/src/components/inGame/waiting/ExitRoom.tsx
